FountainAI-Eco-System/api_gateway/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.23.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
prometheus-fastapi-instrumentator==5.11.2
pytest==7.2.2
pytest-asyncio==0.21.0
